✨(GitHub docs) - In GitHub docs, check if the link can be loaded in an iframe
