+++ src/lib/utils/documentation/getAllDocumentations.ts
@@ -1,53 +1,10 @@
-import { fetch } from "undici";
 import { IDocumentation } from "../../interfaces/IDocumentation";
-import IPackageInformation from "../../interfaces/IPackageInformation";
 import IPackageJson from "../../interfaces/IPackageJson";
+import checkIframeSupport from "../checkIframeSupport";
 import findUrlDocumentation from "../findUrlDocumentation";
 import getFaviconUrl from "../getFaviconUrl";
 import getPackageInfo from "../getPackageInfo";
-
-const formatUrl = async (info: IPackageInformation) => {
-  const url = info.homepage || (info.repository && info.repository.url) || "";
-
-  if (url.includes("radix-ui.com/primitives")) {
-    const componentName = info.name.split("/")[1].split("react-")[1];
-    const componentUrl = `${url}/docs/components/${componentName}`;
-
-    const response = await fetch(componentUrl);
-    if (!response.ok) {
-      const utilityUrl = `${url}/docs/utilities/${componentName}`;
-      const response = await fetch(utilityUrl);
-      if (!response.ok) {
-        return null;
-      }
-      return utilityUrl;
-    }
-
-    return componentUrl;
-  }
-
-  return url;
-};
-
-const checkIframeSupport = async (url: string) => {
-  try {
-    const response = await fetch(url, {
-      method: "HEAD",
-    });
-
-    const xFrameOptions = response.headers.get("X-Frame-Options");
-
-    if (xFrameOptions) {
-      // console.log(`X-Frame-Options: ${xFrameOptions}`);
-      return false;
-    }
-
-    return true;
-  } catch (error) {
-    // console.error("Error fetching the URL:", error);
-    return false;
-  }
-};
+import formatUrl from "./formatUrl";
 
 const getAllDocumentations = async (
   packageJson: IPackageJson,
